+++ src/main/java/com/li64/tide/data/minigame/FishCatchMinigame.java
@@ -61,7 +61,7 @@ public class FishCatchMinigame {
         float fishSpeed = data.map(FishData::speed).orElse(0.5f); // movements per second
 
         // Apply iron fishing hook bonus
-        if (hook.getHook().is(TideItems.IRON_FISHING_HOOK)) fishStrength *= 0.85f;
+        if (hook.getHook().is(TideItems.IRON_HOOK)) fishStrength *= 0.85f;
 
         float area = Mth.clamp(1 - fishStrength, 0.05f, 1.0f);
         float speed = Math.max(fishSpeed / 20f * Tide.CONFIG.minigame.minigameDifficulty, 0.05f);