+++ src/main/java/com/li64/tide/data/rods/FishingRodTooltip.java
@@ -2,4 +2,4 @@ package com.li64.tide.data.rods;
 
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 
-public record FishingRodTooltip(BaitContents contents) implements TooltipComponent {}
+public record FishingRodTooltip(int slots, BaitContents contents) implements TooltipComponent {}